#12
- Refactored code
- Created new class for Alerts
